citrix_metrics/requirements.txt
requests==2.28.2
influxdb-client==1.36.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
tenacity==8.2.2
schedule==1.2.0
python-dotenv==1.0.0
prometheus-client==0.14.1
pyyaml